requirements.txt
torch
torchvision
opencv-python
matplotlib
numpy
pillow
git+https://github.com/facebookresearch/segment-anything.git
